Split up signed in page
