+++ app/javascript/pages/Home/signedIn/BanNotice.svelte
@@ -1,0 +1,38 @@
+<script lang="ts">
+</script>
+
+<div
+  class="text-primary bg-red-500/10 border-2 border-red-500/20 px-5 py-4 rounded-lg mb-4 flex items-start gap-3"
+>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="22"
+    height="22"
+    viewBox="0 0 16 16"
+    class="shrink-0 mt-0.5"
+    ><path
+      fill="currentColor"
+      fill-rule="evenodd"
+      d="M8 14.5a6.5 6.5 0 1 0 0-13a6.5 6.5 0 0 0 0 13M8 16A8 8 0 1 0 8 0a8 8 0 0 0 0 16m1-5a1 1 0 1 1-2 0a1 1 0 0 1 2 0m-.25-6.25a.75.75 0 0 0-1.5 0v3.5a.75.75 0 0 0 1.5 0z"
+      clip-rule="evenodd"
+    /></svg
+  >
+  <div class="text-left">
+    <span class="text-lg font-bold">Your account has been banned.</span>
+    <span class="text-sm opacity-80">
+      Flagged for fraud or abuse - leaderboard and access to Hack Club events
+      may be restricted.
+      <a
+        href="https://fraud.hackclub.com"
+        target="_blank"
+        class="underline font-semibold opacity-100">Learn more</a
+      >
+      ·
+      <a
+        href="https://hackclub.slack.com/team/U091HC53CE8"
+        target="_blank"
+        class="underline font-semibold opacity-100">DM Fraud Dept on Slack</a
+      > (only channel we respond in)
+    </span>
+  </div>
+</div>
